Fix printing of deadline miss ratio.

+++ rt/EarliestDeadlineFirstScheduler.java
@@ -195,6 +195,6 @@ public class EarliestDeadlineFirstScheduler extends Scheduler{
 	
 	private void printDeadlineMissRatio()
 	{
-		System.out.println("Deadline miss ratio: " + (deadlineMisses / finishedTasks));
+		System.out.println("Deadline miss ratio: " + ((float)deadlineMisses / finishedTasks));
 	}
 }
